feat: add functions generateDescription and generatePhotoList
