+++ src/mock/destination.js
@@ -1,0 +1,32 @@
+import {generateRandomRangeNumber} from '../util.js';
+
+const descriptionSentences = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
+  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
+  `Fusce tristique felis at fermentum pharetra.`,
+  `Aliquam id orci ut lectus varius viverra.`,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
+  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
+  `Aliquam erat volutpat.`,
+  `Nunc fermentum tortor ac porta dapibus.`,
+  `In rutrum ac purus sit amet tempus.`
+];
+
+const generatePhotoList = () => {
+  const photos = [];
+  const photoCount = generateRandomRangeNumber(1, 5);
+
+  for (let i = 0; i < photoCount; i++) {
+    photos.push(`http://picsum.photos/300/150?r=${Math.random()}`);
+  }
+
+  return photos;
+};
+
+const generateDescription = (text) => {
+  return text.slice(1, generateRandomRangeNumber(2, 3)).join(` `);
+};
+
+export {generatePhotoList, generateDescription, descriptionSentences};
